requirements.txt
python-telegram-bot==20.0
pytube==12.1.0
requests==2.26.0
python-dotenv==0.19.2